% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_of_life.R
\name{game_of_life}
\alias{game_of_life}
\title{Game of life}
\usage{
game_of_life(sandbox_dimension = 100, iterations = 300,
  probability_of_initial_existence = 0.35, interactive = TRUE,
  delay_time = 10, path = "~/")
}
\arguments{
\item{sandbox_dimension}{(\code{integer}) Dimension of the sandbox,
by default are 100 (a 100 x 100 matrix).}

\item{iterations}{(\code{integer}) Number of simulations that will be stored
in the exported file, by default 300 simulations are stored.}

\item{probability_of_initial_existence}{(\code{double}) Probability that a
cell exists at the beginning of the simulation.}

\item{interactive}{(\code{Boolean}) Should export a .Gif? or return as array?.}

\item{delay_time}{(\code{integer}) Time in milliseconds of waiting for the
frame change in the .GIF animation.}

\item{path}{(\code{string}) Path where the .GIF file will be saved.}
}
\value{
.Gif File with the simulation.
}
\description{
Game of life
}
\examples{
\dontrun{
game_of_life()
game_of_life(sandbox_dimension = 100)
game_of_life(sandbox_dimension = 100, iterations = 300)
game_of_life(sandbox_dimension = 100, iterations = 300,
             probability_of_initial_existence = 0.35)
game_of_life(sandbox_dimension = 100, iterations = 300,
             probability_of_initial_existence = 0.35, delay_time = 5)
game_of_life(sandbox_dimension = 150, iterations = 100,
             probability_of_initial_existence = 0.15, delay_time = 20)
}
}
